# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "싱송생송 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/songs/{songId}/related": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "연관된 노래들과 다음 페이지 번호를 함께 조회합니다. 노래 상세 화면에 첫 진입했을 경우 page 번호는 1입니다. 무한스크롤을 진행한다면 응답에 포함되어 오는 nextPage를 다음번에 포함하여 보내면 됩니다. nextPage는 1씩 증가합니다. 더이상 노래가 없을 경우, 응답에는 빈 배열과 함께 nextPage는 1로 반환됩니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "연관된 노래들을 조회합니다",
                "parameters": [
                    {
                        "type": "string",
                        "description": "songId",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "현재 조회할 노래 목록의 쪽수. 입력하지 않는다면 기본값인 1쪽을 조회",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "한번에 조회할 노래 개수. 입력하지 않는다면 기본값인 20개씩 조회",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.relatedSongResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v2/songs/{songId}/related": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "연관된 노래들과 다음 페이지 번호를 함께 조회합니다. 노래 상세 화면에 첫 진입했을 경우 page 번호는 1입니다. 무한스크롤을 진행한다면 응답에 포함되어 오는 nextPage를 다음번에 포함하여 보내면 됩니다. nextPage는 1씩 증가합니다. 더이상 노래가 없을 경우, 응답에는 빈 배열과 함께 nextPage는 1로 반환됩니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "연관된 노래들을 조회합니다",
                "parameters": [
                    {
                        "type": "string",
                        "description": "songId",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "현재 조회할 노래 목록의 쪽수. 입력하지 않는다면 기본값인 1쪽을 조회",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "한번에 조회할 노래 개수. 입력하지 않는다면 기본값인 20개씩 조회",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/pkg.BaseResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.relatedSongResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.relatedSong": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "isKeep": {
                    "type": "boolean"
                },
                "isLive": {
                    "type": "boolean"
                },
                "isMr": {
                    "type": "boolean"
                },
                "lyricsYoutubeLink": {
                    "type": "string"
                },
                "melonLink": {
                    "type": "string"
                },
                "singerName": {
                    "type": "string"
                },
                "songId": {
                    "type": "integer"
                },
                "songName": {
                    "type": "string"
                },
                "songNumber": {
                    "type": "integer"
                },
                "tjYoutubeLink": {
                    "type": "string"
                }
            }
        },
        "handler.relatedSongResponse": {
            "type": "object",
            "properties": {
                "nextPage": {
                    "type": "integer"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.relatedSong"
                    }
                }
            }
        },
...
        "pkg.BaseResponseStruct": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}